backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
babel==2.17.0
Beaker==1.12.1
beautifulsoup4==4.14.2
black==25.11.0
bleach==6.2.0
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.40.61
botocore==1.40.61
Brlapi==0.8.6
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-347-build/cockpit-347/tmp/wheel/cockpit-347-py3-none-any.whl#sha256=8816ee14baaafcd61279d6180ad0c57fd497ac6ad5d73db9bc0de1fa66d9e21d
comm==0.2.3
contourpy==1.3.3
crypt_r==3.13.1
cryptography==44.0.0
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnf==4.24.0
executing==2.2.0
fastapi==0.120.0
fastjsonschema==2.21.2
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.20.0
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.56.0
fqdn==1.5.1
fros==1.1
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanize==4.12.0
i3ipc==2.2.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inkex==1.4.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
iso639==0.1.4
isoduration==20.11.0
isort==7.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.22
libdnf==0.75.0
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mccabe==0.7.0
mercurial==6.9.5
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.2
mutagen==1.47.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.5
nftables==0.1
nltk==3.9.2
nodeenv==1.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
nwg-displays==0.3.26
nwg-panel==0.10.12
olefile==0.47
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
Paste==3.10.1
pathspec==0.12.1
pdfminer.six==20251107
pdfplumber==0.11.8
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.2.2
ply==3.11
pooch==1.8.2
productmd==1.45
progressbar2==4.5.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.25.1
pycodestyle==2.14.0
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.23.0
pycups==2.0.4
pydantic==2.12.3
pydantic_core==2.41.4
pyenchant==3.2.2
pyflakes==3.4.0
Pygments==2.19.2
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
pynvim==0.6.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
pyparted==3.13.0
pypdfium2==5.0.0
pyphen==0.17.2
pyquery==2.0.1
pyright==1.1.407
pyserial==3.5
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.8.2
python-docx==1.2.0
python-json-logger==3.3.0
python-linux-procfs==0.7.3
python-meh==0.52
python-multipart==0.0.20
python-pam==2.0.2
python-utils==3.9.0
python-xlib==0.33
pytokens==0.3.0
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.17.6-build/kernel-6.17.6/linux-6.17.6-200.fc42.x86_64/tools/net/ynl
pyzmq==27.0.1
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.0
rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.14.1
scour==0.38.2
screeninfo==0.8.1
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
Send2Trash==1.8.3
sentence-transformers==5.1.1
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
systemd-python==235
terminado==0.18.1
textstat==0.7.11
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
triton==3.4.0
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.35.0
waypaper==2.7
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
xkbregistry==0.3
yt-dlp==2025.10.22
